move ROI-class
