+++ fmriproc/image.py
@@ -875,138 +875,6 @@ def tsnr(img,file_name=None, clip=None):
 
     return mean_tsnr
 
-class ROI():
-
-    def __init__(
-        self,
-        roi,
-        mask_type="aparc",
-        subject="fsaverage",
-        ):
-
-        self.roi = roi
-        self.mask_type = mask_type
-        self.subject = subject
-
-        # force list
-        if isinstance(self.roi, str):
-            self.roi = [self.roi]
-
-        # make mask from aparc segmentation
-        if self.mask_type == "aparc":
-            self.aparc_mask()
-        else:
-            # make mask from surf-labels
-            self.surf_mask()
-
-        # merge list of masks
-        self.merge_masks()
-        
-    def surf_mask(self):
-
-        try:
-            from cxutils import optimal
-        except ImportError:
-            print("Could not import cxutils. Please install from https://github.com/gjheij/cxutils")
-
-        self.surf_calcs = optimal.SurfaceCalc(subject=self.subject)
-
-        # loop through list and merge
-        self.individual_masks = {}
-        if isinstance(self.roi, list):
-            for roi in self.roi:
-
-                self.mask_obj = self.surf_calcs.read_fs_label(
-                    self.subject, 
-                    fs_label=roi
-                )
-                
-                self.individual_masks[roi] = self.surf_calcs.label_to_mask(
-                    subject=self.subject, 
-                    rh_arr=self.mask_obj['rh'],
-                    lh_arr=self.mask_obj['lh']
-                )["whole_roi"]
-
-    def aparc_mask(self):
-
-        # read parc data once even for multiple ROIs
-        self.read_aparc()
-        self.roi_list = self.read_roi_list()
-
-        # loop through list and merge
-        self.individual_masks = {}
-        if isinstance(self.roi, list):
-            for roi in self.roi:
-                if roi not in self.roi_list:
-                    raise ValueError(f"'{roi}' is not part of the aparc atlas. Options are {self.roi_list}")
-                
-                self.individual_masks[roi] = self.make_roi_mask(roi=roi)
-
-    def merge_masks(self):
-
-        # merge
-        if len(self.individual_masks)>1:
-            self.roi_mask = np.zeros_like(self.individual_masks[self.roi[0]])
-            for _,val in self.individual_masks.items():
-                self.roi_mask[val>0] = 1
-        else:
-            self.roi_mask = self.individual_masks[self.roi[0]].copy()
-
-    def get_rois(self):
-        return self.read_roi_list()
-    
-    def return_mask(self):
-        return self.roi_mask
-    
-    def read_aparc(self):
-
-        try:
-            from cxutils import optimal
-        except ImportError:
-            print("Could not import cxutils. Please install from https://github.com/gjheij/cxutils")        
-
-        # GET VERICES FOR A SPECIFIC ROI 
-        self.parc_data = optimal.SurfaceCalc.read_fs_annot(
-            subject='fsaverage',
-            fs_annot="aparc",
-            hemi="both"
-        )
-
-    def read_roi_list(self):
-        if not hasattr(self, "parc_data"):
-            self.read_aparc()
-
-        roi_list = []
-        for i in self.parc_data["lh"][-1]:
-            roi_list.append(i.decode())
-
-        roi_list = [i for i in roi_list if i != "unknown"]
-        return roi_list
-    
-    def make_roi_mask(self, roi=None):
-        
-        if not isinstance(roi, str):
-            raise ValueError(f"Please specify an ROI to extract")
-
-
-        tmp = {}
-        for ii in ['lh', 'rh']:
-            
-            # get data
-            parc_hemi = self.parc_data[ii][0]
-
-            # get index of specified ROI in list | hemi doesn't matter here
-            for ix,i in enumerate(self.parc_data[ii][2]):
-                if roi.encode() in i:
-                    break
-            
-            tmp[ii] = (parc_hemi == ix).astype(int)
-
-        # Concat Hemis
-        concat_hemis = np.hstack((tmp['lh'],tmp['rh']))
-
-        return concat_hemis
-
 def massp_to_table(label_file, out=None, nr_structures=31, unit="vox"):
     """massp_to_table
